Time: 149 ms (88.17%), Space: 11.4 MB (19.33%) - LeetHub

+++ 0204-count-primes/0204-count-primes.cpp
@@ -4,11 +4,11 @@ public:
         int cnt=0;
         bool prime[n+1];
         memset(prime,0,sizeof(prime));
-        for(int i=2;i*i<=n-1;i++)
+        for(long long i=2;i*i<=n-1;i++)
         {
             if(prime[i]==0)
             {
-                for(int j=i*i;j<=n-1;j+=i)
+                for(long long j=i*i;j<=n-1;j+=i)
                 {
                     if(prime[j]==0)
                     {
@@ -17,7 +17,7 @@ public:
                 }
             }
         }
-        for(int i=2;i<=n-1;i++)
+        for(long long i=2;i<=n-1;i++)
         {
             if(prime[i]==0)
             {
